refactor: simplify resource management by moving logic from TurnAndAiService to GameApi
