+++ .github/copilot-instructions.md
@@ -12,7 +12,7 @@
 
 ### TDD 開発プロセス (標準フロー)
 
-すべての機能開発は以下の 4 ステップで進める。
+すべての機能開発は以下のステップで進める。
 
 1. **[人間]** 仕様の入力 — 自然言語または copilot-instructions.md / GameApi TSDoc への追記。
 2. **[AI]** テストケースファイルの作成 — テストファイルを作成し、`describe` / `test` の構造と名前のみを定義。テスト関数の中身は空にする。
