+++ .github/workflows/ci.yml
@@ -27,9 +27,6 @@ jobs:
       - name: Install dependencies
         run: pnpm install
 
-      - name: Format check
-        run: pnpm prettier --check .
-
       - name: Lint
         run: pnpm lint
